requirements.txt
flask==3.0.0
requests==2.31.0
urllib3==2.0.7
python-dotenv==1.0.0
pytest==7.4.3

